Remove advice that ended up being unnecessary.
A while back I had problems with Emacs loading the wrong org-mode when it
started up, and one suggested fix was to wrap package-installed-p with some
advice that returns t only for packages that were actually installed through
package.el. The problem was that a call to (package-installed-p 'emacs) return
nil which caused version checks against emacs to fail and hence package
refusing to install any package that had emacs listed in its dependencies.

As it turns out, the advice isn't needed if you add org to the
package-pinned-packages alist prior to calling package-initialize. At least
that seems to be the case. Regardless, removing the advice allowed other
packages (yaml-mode and plantuml-mode) to install. Huzzah.